feat: centered up and down title when no content in custom container
When there's no content in custom container, the title of custom container is not centered up and down, and the reason is that `margin-bottom: -0.4em;` overwrites the original `margin-bottom` of the `p` tag. So removing this style ensures that the title is center when no content in custom container. In addition, the container is not affected when it has content.