Escape double-quotes in head tag attribute values
To reproduce the bug:

```
module.exports = {
  head: [
    ['script', {
       src: 'https://something.com/foo.js',
       'data-payload': '{"token": "hello"}',  // Use an attribute value with a JSON-encoded string.
    }],
  ],
```

You should run into the following error:

```
Document.querySelectorAll: 'head > script[src="https://something.com/foo.js"][data-payload="{"token": "hello"}"]' is not a valid selector
```

The problem is that the `data-payload` value is a JSON-encoded string, which requires double quotes, but the current implementation simply wraps the CSS selector for that call to `querySelectorAll` in double quotes and calls it a day. Use `JSON.stringify` instead. Since `value` is known to be a string in this context, this should always work (tested locally manually).